sdsd
sdsd

+++ app/views/home/welcome.blade.php
@@ -141,7 +141,20 @@
                                   <h5>{{$file->FECHA}}</h5>  
                                   
                                   </a> 
-                                  <center><a href="eliminar/{{$file->NOMBRE}}" title="Eliminar archivo: {{$file->NOMBRE}} "> <p style="color: red;" class="glyphicon glyphicon-remove"> </p> </a></center>
+
+                                  <script language="Javascript">
+                                       function confirmDel()
+                                        {
+                                          var agree= smoke.confirm("¿Realmente desea eliminar el archivo? ");
+                                          if (agree){
+                                            return true ;
+                                          }
+                                          else{
+                                             return false ;}
+                                        }
+                                  </script>
+                                    <!-- Eliminar archivo -->
+                                   <center><a onclick="return confirmDel();"  href="eliminar/{{$file->NOMBRE}}" title="Eliminar archivo: {{$file->NOMBRE}} "> <p style="color: red;" class="glyphicon glyphicon-remove"> </p> </a></center>
         
                                    @endforeach
                                    @endif
